move r_matrix calculation to elements
