+++ ocelot/cpbd/response_matrix.py
@@ -1,6 +1,7 @@
 from ocelot.cpbd.optics import *
 from ocelot.cpbd.match import closed_orbit
 from ocelot.cpbd.track import tracking_step
+from ocelot.cpbd.elements import *
 import copy
 import os
 import numpy as np
